Spawns and 1 byte bool

+++ move.h
@@ -1,14 +1,15 @@
 #ifndef CHESS_MOVE_H
 #define CHESS_MOVE_H
 
-#include <stdbool.h>
+#include <stdint.h>
+#include "defs.h"
 
 typedef struct Move {
-    int x, y;
+    int8_t x, y;
     bool repeat;
 } Move;
 
-void initMove(Move * move, int x, int y, bool repeat);
+void initMove(Move * move, int8_t x, int8_t y, bool repeat);
 void printMove(Move move);
 
 #endif
